Add a script to run the parallel application.
We need to know the number of frames for the video to be
proceesed. This is needed because our processes changes
messages in a loop until all the frames are processed.

In the implementation master read a frame and send chunks of
data to the other processes. For a good flow, both master and
processes will send and receive data in a loop(frames number)